starting point derived from the functional original version. 2025-02-20.
The VS project files are included and assume the existence of PDCursesMod
which has project files but no content. It is realized only in the original project.
One of the targets is to remove it and replace it by a reasonable GUI implementation.

The examples file typically used for tests comes from the online available windows version.